Move service data deletion logic to connected_services module
The service data download logic was already moved to connected_service
module before (8bdad794de3cbd5bfe713e9f3e9491bb03ad518b). Now the
deletion logic is moved to the same place.
